add a checksum for the source
The benefits of this change outweigh the cons. The download will now be
checked and for the cases where there is any chance of a rebuild the
cached download will be used.

The drawback is that source-checksum does not support advanced grammar,
this should be minor considering the build-on restriction for the amd64
as an architecture to target.

Signed-off-by: Sergio Schvezov <sergio.schvezov@ubuntu.com>